+++ lib/mix/tasks/compile.build_dot_zig.ex
@@ -25,14 +25,25 @@ defmodule Mix.Tasks.Compile.BuildDotZig do
     downloaded and installed locally in the build directory. If `false`, no Zig installation is
     downloaded. If `true` (default), the latest Zig stable version is downloaded. Otherwise, it's
     possible to pass a specific Zig version, e.g. `install_zig: "0.9.1"`.
+
     * `:zig_executable` - (binary or `:default`) it's the executable to use as the `zig`
     program. If not provided or if `:default`, it defaults to the downloaded `zig` binary if
     the `install_zig` was configured to download Zig, otherwise it defaults to `zig` (it assumes
     to find it in the `PATH`). Note that it's possible to both install a downloaded Zig installation
     _and_ pass a local `:zig_executable`. This is useful for use cases where the downloaded Zig
     installation must be called through some wrapper script.
+
     * `:zig_build_mode` - (atom) allows choosing the build mode. The supported build modes are
     `:debug` (default), `:release_safe`, `:release_fast` and `:release_small`.
+
+    * `:zig_target` - (binary or `:host`) it's the target that will be passed with the
+    `-Dtarget` flag to `zig`. This can be used to support cross-compilation. If not provided or if
+    `:host` (which is the default returned by `Mix.target()`), the `-Dtarget` flag is not passed.
+
+    * `:zig_cpu` - (binary or `:native`) the set of CPU features that will be passed with the
+    `-Dcpu` flag to `zig`. If not provided or if `:native` the `-Dcpu` flag is not passed, which
+    makes `zig` use _all_ the CPU features of the host. If you're not running the compiled code on
+    the same machine where you're compiling, setting this to `"baseline"` is a good starting point.
 
   ## Default environment variables