+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.3.0] - Unreleased
 
+### Added
+
+- Add `zig_target` option.
+- Add `zig_cpu` option.
+
 ## [0.2.0] - 2023-07-10
 
 ### Added